add separate nn for nirc to learn gi and nee with two different nn
A single nn cannot learn gi and nee at the same time as they will often
give contradicting gradients. In pt this is no problem as mis takes care
of that. However, in gradients one gradient will dominate and the nn
will learn the dominating one. This means that nee leads to black output
of the nn at points where it tries to connect to a light source but
cannot connect. At those points the indirect bounce will find light but
the gradient will be dominated by the black nee.